pyproject.toml
[project]
name = "attestationwork"
version = "0.1.0"
description = ""
authors = [
    {name = "Sepega",email = "twnue2020@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "eventlet (>=0.40.3,<0.41.0)",
    "celery (>=5.5.3,<6.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "redis (>=7.0.1,<8.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
